move extraction functions to graph classes
