+++ tests/test_countgraph.py
@@ -116,6 +116,38 @@ def test_revhash_1():
     assert hi.reverse_hash(hashval) == kmer
 
 
+def test_extract_countgraph_info_badfile():
+    try:
+        Countgraph.extract_info(
+            utils.get_test_data('test-abund-read-2.fa'))
+        assert 0, 'this should fail'
+    except ValueError:
+        pass
+
+
+def test_extract_countgraph_info():
+    fn = utils.get_temp_filename('test_extract_counting.ct')
+    for size in [1e6, 2e6, 5e6, 1e7]:
+        ht = khmer.Countgraph(25, size, 4)
+        ht.save(fn)
+
+        try:
+            info = Countgraph.extract_info(fn)
+        except ValueError as err:
+            assert 0, 'Should not throw a ValueErorr: ' + str(err)
+        ksize, n_tables, table_size, _, _, _, _ = info
+        print(ksize, table_size, n_tables)
+
+        assert(ksize) == 25
+        assert table_size == size
+        assert n_tables == 4
+
+        try:
+            os.remove(fn)
+        except OSError as err:
+            assert 0, '...failed to remove ' + fn + str(err)
+
+
 class Test_Countgraph(object):
 
     def setup(self):
