+++ tests/test_nodegraph.py
@@ -34,9 +34,6 @@
 # Contact: khmer-project@idyll.org
 # pylint: disable=missing-docstring,protected-access,no-member,invalid-name
 
-from __future__ import print_function
-from __future__ import absolute_import
-
 import khmer
 from khmer import Nodegraph, Countgraph
 from khmer import ReadParser
@@ -46,6 +43,7 @@ from khmer.khmer_args import create_matching_nodegraph
 import screed
 
 import pytest
+import os
 
 from . import khmer_tst_utils as utils
 
@@ -61,6 +59,36 @@ def test_toobig():
         assert 0, "This should fail"
     except MemoryError as err:
         print(str(err))
+
+
+def test_extract_nodegraph_info_badfile():
+    try:
+        Nodegraph.extract_info(
+            utils.get_test_data('test-abund-read-2.fa'))
+        assert 0, 'this should fail'
+    except ValueError:
+        pass
+
+
+def test_extract_nodegraph_info():
+    fn = utils.get_temp_filename('test_extract_nodegraph.pt')
+    for size in [1e6, 2e6, 5e6, 1e7]:
+        ht = khmer.Nodegraph(25, size, 4)
+        ht.save(fn)
+
+        info = Nodegraph.extract_info(fn)
+        ksize, table_size, n_tables, _, _, _ = info
+        print(ksize, table_size, n_tables)
+
+        assert(ksize) == 25
+        assert table_size == size, table_size
+        assert n_tables == 4
+
+        try:
+            os.remove(fn)
+        except OSError as err:
+            print('...failed to remove {fn}'.format(fn) + str(err),
+                  file=sys.stderr)
 
 
 def test_add_tag():
